Ignore the `anonymous` user that exists on SteamOS
SteamOS will add a folder to the userdata directory named `anonymous` by default. Previously this folder only contained folders where the folder name was the userid of a user that had logged in to Steam, so I navively casted the directory name to an integer. This caused a crash when ran on SteamOS because of this anonymous user.

This commit makes the `local_users` function ignore any directory name that can't be casted to an integer.

Fixes #2