Rewrite the API for v1.0
I've learned a lot since I originally made the pysteam API, so hopefully this one is easier to work with (and easier to test with!). One of my main goals here was to separate data and functionality. As I'm gearing up to write integration tests for Ice I've found pysteam to be a big problem. Since its an integration test I dont want to mock anything, but I also don't want it actually using the data on my local machine when running tests. That just makes them nondeterministic.

Now, my value object classes are all stored in `model.py`, and all of my functionality exists as pure functions in an appropriate namespace that take the necessary data as parameters. This means I need to do significantly less mocking - now I just have to send in the right data!

Also move all of the old code to `legacy/`, in case some poor soul actually still wants to use the API. I will probably remove it at a later date.